音视频处理，获取时长，封面截图

+++ src/main/java/com/imageserver/image/bean/VideoBean.java
@@ -26,7 +26,7 @@ public class VideoBean implements Serializable {
     @ApiModelProperty(value = "视频封面地址", name = "frontCover")
     private String frontCover;
     @ApiModelProperty(value = "时长", name = "duration")
-    private double duration;
+    private Long duration;
 
 
 }
